+++ canaille/core/endpoints/account.py
@@ -36,15 +36,14 @@ from canaille.app.flask import user_needed
 from canaille.app.forms import IDToModel
 from canaille.app.forms import TableForm
 from canaille.app.forms import is_readonly
+from canaille.app.forms import password_too_long_validator
+from canaille.app.forms import pwned_password_validator
 from canaille.app.forms import set_readonly
 from canaille.app.forms import set_writable
 from canaille.app.i18n import gettext as _
 from canaille.app.i18n import reload_translations
 from canaille.app.themes import render_template
 from canaille.backends import Backend
-from canaille.app.forms import password_strength_calculator
-from canaille.app.forms import password_too_long_validator
-from canaille.app.forms import pwned_password_validator
 
 from ..mails import send_confirmation_email
 from ..mails import send_invitation_mail
@@ -324,9 +323,6 @@ def registration(data=None, hash=None):
     ]
     form["password2"].validators = [
         wtforms.validators.DataRequired(),
-        wtforms.validators.Length(
-            min=current_app.config["CANAILLE"]["MIN_PASSWORD_LENGTH"]
-        ),
     ]
     form["password1"].flags.required = True
     form["password2"].flags.required = True
@@ -661,7 +657,7 @@ def profile_edition(user, edited_user):
 
 @bp.route("/profile/<user:edited_user>/settings", methods=("GET", "POST"))
 @user_needed()
-def profile_settings(user, edited_user, password_strength=0):
+def profile_settings(user, edited_user):
     if not user.can_manage_users and not (
         user.can_edit_self and edited_user.id == user.id
     ):
@@ -672,7 +668,7 @@ def profile_settings(user, edited_user, password_strength=0):
         or request.form.get("action") == "edit-settings"
         or request_is_htmx()
     ):
-        return profile_settings_edit(user, edited_user, request.args.get("pwds", 0))
+        return profile_settings_edit(user, edited_user)
 
     if request.form.get("action") == "confirm-delete":
         return render_template("modals/delete-account.html", edited_user=edited_user)
@@ -749,7 +745,7 @@ def profile_settings(user, edited_user, password_strength=0):
     abort(400, f"bad form action: {request.form.get('action')}")
 
 
-def profile_settings_edit(editor, edited_user, password_strength=0):
+def profile_settings_edit(editor, edited_user):
     menuitem = "profile" if editor.id == editor.id else "users"
     fields = editor.readable_fields | editor.writable_fields
 
@@ -775,7 +771,6 @@ def profile_settings_edit(editor, edited_user, password_strength=0):
         and request.form.get("action") == "edit-settings"
         or request_is_htmx()
     ):
-        password_strength = password_strength_calculator(form["password1"].data)
         if not form.validate():
             flash(_("Profile edition failed."), "error")
 
@@ -794,7 +789,7 @@ def profile_settings_edit(editor, edited_user, password_strength=0):
             Backend.instance.save(edited_user)
             flash(_("Profile updated successfully."), "success")
             return redirect(
-                url_for("core.account.profile_settings", edited_user=edited_user, pwds=password_strength)
+                url_for("core.account.profile_settings", edited_user=edited_user)
             )
 
     return render_template(
@@ -803,7 +798,6 @@ def profile_settings_edit(editor, edited_user, password_strength=0):
         menuitem=menuitem,
         edited_user=edited_user,
         self_deletion=edited_user.can_delete_account,
-        password_strength=password_strength
     )