+++ tests/core/test_profile_settings.py
@@ -1,8 +1,8 @@
 import datetime
 from unittest import mock
 
-from flask import g
 from flask import current_app
+from flask import g
 
 from canaille.app import models
 
@@ -94,11 +94,12 @@ def test_profile_settings_edition_dynamic_validation(testclient, logged_admin):
             "HX-Trigger-Name": "password1",
         },
     )
-    res.mustcontain("Password is too short, minimum length: 8")
+    res.mustcontain("Field must be at least 8 characters long.")
+
 
 def test_profile_settings_minimum_password_length_validation(testclient, logged_user):
-    """Tests minimum length of password defined in configuration.
-    """
+    """Tests minimum length of password defined in configuration."""
+
     def with_different_values(password, length):
         current_app.config["CANAILLE"]["MIN_PASSWORD_LENGTH"] = length
         res = testclient.get("/profile/user/settings")
@@ -113,16 +114,18 @@ def test_profile_settings_minimum_password_length_validation(testclient, logged_
                 "HX-Trigger-Name": "password1",
             },
         )
-        res.mustcontain(f"Password is too short, minimum length: {length}")
+        res.mustcontain(f"Field must be at least {length} characters long.")
 
     with_different_values("short", 8)
     with_different_values("aa", 3)
     with_different_values("1234567890123456789", 20)
 
+
 def test_profile_settings_too_long_password(testclient, logged_user):
-    """Tests maximum length of password.
-    """
-    def with_different_values(password, message):
+    """Tests maximum length of password."""
+
+    def with_different_values(password, length, message):
+        current_app.config["CANAILLE"]["MAX_PASSWORD_LENGTH"] = length
         res = testclient.get("/profile/user/settings")
         res = testclient.post(
             "/profile/user/settings",
@@ -137,13 +140,21 @@ def test_profile_settings_too_long_password(testclient, logged_user):
         )
         res.mustcontain(message)
 
-    with_different_values("a"*1003, "Invalid password")
-    with_different_values("a"*1001, 'password_strength="100"')
+    with_different_values(
+        "a" * 1001, 1000, "Field cannot be longer than 1000 characters."
+    )
+    with_different_values("a" * 1000, 1000, 'data-percent="100"')
+    with_different_values("a" * 501, 500, "Field cannot be longer than 500 characters.")
+    with_different_values("a" * 500, 500, 'data-percent="100"')
+
 
 def test_profile_settings_pwned_password(testclient, logged_user):
     """Tests if password is compromised.
-        this function only works with password length > min_length_value (default = 8)
+
+    this function only works with password length > min_length_value
+    (default = 8)
     """
+
     def with_different_values(password, message):
         res = testclient.get("/profile/user/settings")
         res = testclient.post(
@@ -161,8 +172,9 @@ def test_profile_settings_pwned_password(testclient, logged_user):
 
     with_different_values("aaaaaaaa", "This password is compromised.")
     with_different_values("azertyuiop", "This password is compromised.")
-    with_different_values("a"*1001, 'password_strength="100"')
+    with_different_values("a" * 1001, 'password_strength="100"')
     with_different_values("i'm a little pea", 'password_strength="28"')
+
 
 def test_edition_without_groups(
     testclient,